Fixed bug in order to use SSL/TLS over Socks5
It seems that ssl module creates a new socket (ssl.SSLSocket) which uses the method socket.socket instead of socks.socksocket if the socket is not connected before wrapping it. In this way, now we can use the Socks5 proxy properly because before this update the socket connected directly to the server.